Add custom Credo configuration
This is the same as the current default Credo configuration that Ebert
uses, with one change: we've disabled the `SpaceAroundOperators` check
because it flags bindings in function heads, which we tend to write
Erlang-style.

That is, we write like this:

```elixir
def handle_call(msg, from, %__MODULE__{foo: foo}=state)
```

vs. this:

```elixir
def handle_call(msg, from, %__MODULE__{foo: foo} = state)
```

But we put spaces around the `=` everywhere else.